BackendAPI/requirements.txt
alabaster==0.7.9
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astroid==1.4.7
astropy==1.2.1
awsebcli==3.10.5
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
beautifulsoup4==4.5.1
bitarray==0.8.1
blaze==0.10.1
bleach==2.0.0
blessed==1.14.2
bokeh==0.12.2
boto==2.42.0
boto3==1.4.6
botocore==1.6.2
Bottleneck==1.1.0
cement==2.8.2
certifi==2017.7.27.1
cffi==1.7.0
chardet==3.0.4
chest==0.2.3
click==6.6
cloudpickle==0.2.1
colorama==0.3.7
configobj==5.0.6
contextlib2==0.5.3
cryptography==1.5
cycler==0.10.0
Cython==0.24.1
cytoolz==0.8.0
dask==0.11.0
datashape==0.5.2
decorator==4.0.10
defusedxml==0.5.0
dill==0.2.5
Django==1.11.4
django-braces==1.11.0
django-crispy-forms==1.6.1
django-extensions==1.8.1
django-filter==1.0.3
django-markdown-deux==1.0.5
django-oauth-toolkit==1.0.0
django-pagedown==0.1.3
django-rest-framework-social-oauth2==1.0.8
django-social-auth==0.7.28
django-storages==1.6.3
django-warrant==0.1.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
ecdsa==0.13
entrypoints==0.2.3
envs==1.2.2
et-xmlfile==1.0.1
Faker==0.7.17
fastcache==1.0.2
filelock==2.0.6
Flask==0.11.1
Flask-Cors==2.1.2
future==0.16.0
gevent==1.1.2
greenlet==0.4.10
h5py==2.6.0
HeapDict==1.0.0
html5lib==0.999999999
httpie==0.9.9
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jmespath==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
lazy-object-proxy==1.2.1
llvmlite==0.19.0
locket==0.2.0
lxml==3.6.4
Markdown==2.6.8
markdown2==2.3.4
MarkupSafe==0.23
mistune==0.7.3
mock==2.0.0
mpmath==0.19
multipledispatch==0.4.8
nbconvert==4.2.0
nbformat==4.1.0
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
numba==0.28.1
numexpr==2.6.1
numpy==1.11.1
oauth2==1.9.0.post1
oauth2-provider==0.0
oauthlib==2.0.2
odo==0.5.0
openpyxl==2.3.2
pandas==0.18.1
pandocfilters==1.4.2
partd==0.3.6
pathlib2==2.1.0
pathspec==0.5.0
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.1.0
pkginfo==1.3.2
ply==3.9
prompt-toolkit==1.0.3
psutil==4.3.1
psycopg2==2.7.3
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pyflakes==1.3.0
Pygments==2.1.3
PyJWT==1.5.2
pylint==1.5.4
PyMySQL==0.7.11
pyOpenSSL==16.2.0
pyparsing==2.1.4
pytest==2.9.2
python-dateutil==2.5.3
python-decouple==3.1
python-jose==1.3.2
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
pyzmq==15.4.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
rope-py3k==0.9.4.post1
s3transfer==0.1.10
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.1
semantic-version==2.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sockjs-tornado==1.0.3
Sphinx==1.4.6
spyder==3.0.0
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.0
tabulate==0.7.5
termcolor==1.1.0
terminado==0.6
terminaltables==3.1.0
testpath==0.3.1
toolz==0.8.0
tornado==4.4.1
traitlets==4.3.0
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
warrant==0.2.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
Werkzeug==0.11.11
widgetsnbextension==1.2.6
wrapt==1.10.6
xlrd==1.0.0
XlsxWriter==0.9.3
xlwings==0.10.0
xlwt==1.1.2
